add secure validation

+++ src/main/java/com/softserve/edu/bookinglite/service/PhotoService.java
@@ -4,12 +4,14 @@ import java.util.Map;
 import com.cloudinary.*;
 import com.cloudinary.utils.ObjectUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.event.ApplicationEventMulticaster;
 import org.springframework.core.task.TaskExecutor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import com.softserve.edu.bookinglite.config.CloudinaryConfig.UploadOptions;
 import com.softserve.edu.bookinglite.entity.Photo;
+import com.softserve.edu.bookinglite.events.UploadPhotoEvent;
 import com.softserve.edu.bookinglite.repository.PhotoRepository;
 import com.softserve.edu.bookinglite.repository.PropertyRepository;
 @Service
@@ -20,43 +22,39 @@ public class PhotoService {
 	private PropertyRepository propertyRepository;
 	@Autowired
 	private Cloudinary cloudinary;
-	@Autowired  
-	private TaskExecutor taskExecutor;
+	@Autowired
+	private ApplicationEventMulticaster applicationEventMulticaster;
 	
-	private final String URL_RESULT_KEY = "url";
 	private final int MAX_BYTES = 1024*1024*15;
 	private final String AVAILABLE_TYPE = "image";
-	@Async
-	public boolean uploadPhoto(MultipartFile file, Long property_id) {
+	
+	public boolean uploadPhoto(MultipartFile file, Long property_id, Long user_id) {
 		
-		if(file==null || file.getContentType().indexOf(AVAILABLE_TYPE)==-1 || file.getSize()>MAX_BYTES) {
+		
+		if(!propertyRepository.getOne(property_id).getUser().getId().equals(user_id) 
+				|| file==null 
+				|| file.getContentType().indexOf(AVAILABLE_TYPE)==-1 
+				|| file.getSize()>MAX_BYTES) {
 			return false;
+		}else {
+			
+			try {
+				applicationEventMulticaster.multicastEvent(new UploadPhotoEvent(file.getBytes(), property_id));
+			} catch (IOException e) {
+				e.printStackTrace();
+				return false;
+			}
 		}
-		
-		taskExecutor.execute(new Runnable() {  
-			   @Override  
-			   public void run() {  
-					Map options = ObjectUtils.asMap(                 
-							UploadOptions.NAME_OPTION.getName(), UploadOptions.NAME_OPTION.getValue(),
-							UploadOptions.FOLDER_OPTION.getName(), UploadOptions.FOLDER_OPTION.getValue());
-					Map uploadResult=null;
-					try {
-						uploadResult = cloudinary.uploader().upload(file.getBytes(), options);
-					} catch (IOException e) {
-						e.printStackTrace();
-					}
-					Photo photo = new Photo();
-					photo.setProperty(propertyRepository.getOne(property_id));
-					photo.setUrl((String)uploadResult.get(URL_RESULT_KEY));
-					photoRepository.save(photo);   
-			   }  
-			 });  
 		return true;
 	}
-	//name example: 047059ef-6950-469c-8d44-90a311f39982
-	public boolean deletePtoto(String name) {
+	
+	public boolean deletePtoto(String name, Long user_id) {
 		try {
 			Photo photo = photoRepository.findByUrlLike(name).get(0);
+			
+			if(!photo.getProperty().getUser().getId().equals(user_id))
+				throw new Exception("not secure");
+			
 			cloudinary.uploader().destroy(UploadOptions.FOLDER_OPTION.getValue()+"/"+name, ObjectUtils.emptyMap());
 			photoRepository.delete(photo);
 		}catch(Exception exc) {
